Move data generation files into the database folder
