+++ server/database/generateLargeData.js
@@ -2,11 +2,12 @@ const faker = require('faker');
 const fs = require('fs');
 
 /*
-SCHEMA:
+POSTGRESQL SCHEMA:
 TABLE 1: products
-ProductId (unique integer) -- i
-ProductName (string) -- faker.commerce.productName()
-isLiked (boolean)
+productId (unique integer) -- i
+productName (string) -- faker.commerce.productName()
+productSeller (string) -- faker.internet.userName()
+isLiked (boolean) -- getIfLiked()
 
 TABLE 2: product_images
 imageId (unique integer) -- i
@@ -14,18 +15,21 @@ imageUrl (string) -- faker.image.imageUrl()
 ProductId (Foreign Key - integer) -- getRandomProductId(10M)
 */
 
-const getRandomNumber = (max) => {
-  return Math.ceil(Math.random() * max);
-};
+const getIfLiked = () => {
+  let value = getRandomNumber(10);
+  if (value <= 2) {
+    return true;
+  }
+  return false;
+}
 
-//create the stream to write to csv file
+//create the stream to write to the csv file
 const productDataCsv = fs.createWriteStream('./productData.csv')
 
-
-function generateProductDataCsv(total, callback) {
+function generateProductDataCsv(total) {
   let i = total;
   const generateProductLine = () => {
-    return `${i},${faker.commerce.productName()},false\n`;
+    return `${i},${faker.commerce.productName()},${faker.internet.userName()},${getIfLiked()}\n`;
   }
 
   write();
@@ -35,14 +39,12 @@ function generateProductDataCsv(total, callback) {
     do {
       i--;
       ok = productDataCsv.write(generateProductLine(), 'utf8');
-    } while (i > 1 && ok);
+    } while (i > 0 && ok);
     if (i > 0) {
       productDataCsv.once('drain', write);
-    } else if (i === 0) {
-      productDataCsv.write(generateProductLine(), 'utf8', callback);
     }
   }
 }
 
-generateProductDataCsv(10000000, console.log);
+generateProductDataCsv(10000000);
